Remove templates from Zarr object.
