+++ src/zarr.blosc.cpp
@@ -6,6 +6,7 @@
 #include <stdexcept>
 
 namespace zarr = acquire::sink::zarr;
+using json = nlohmann::json;
 
 namespace {
 template<zarr::BloscCodecId CodecId>
@@ -13,7 +14,9 @@ struct Storage*
 compressed_zarr_init()
 {
     try {
-        return new zarr::Zarr<zarr::BloscEncoder<CodecId, 1, 1>>();
+        zarr::BloscCompressor compressor(
+          zarr::compression_codec_as_string<CodecId>(), 1, 1);
+        return new zarr::Zarr(std::move(compressor));
     } catch (const std::exception& exc) {
         LOGE("Exception: %s\n", exc.what());
     } catch (...) {
@@ -22,6 +25,67 @@ compressed_zarr_init()
     return nullptr;
 }
 } // end ::{anonymous} namespace
+
+//
+// zarr namespace implementations
+//
+
+void
+zarr::to_json(json& j, const zarr::BloscCompressor& bc)
+{
+    j = json{ { "id", std::string(bc.id_) },
+              { "cname", bc.codec_id_ },
+              { "clevel", bc.clevel_ },
+              { "shuffle", bc.shuffle_ } };
+}
+
+void
+zarr::from_json(const json& j, zarr::BloscCompressor& bc)
+{
+    j.at("cname").get_to(bc.codec_id_);
+    j.at("clevel").get_to(bc.clevel_);
+    j.at("shuffle").get_to(bc.shuffle_);
+}
+
+zarr::BloscEncoder::BloscEncoder(const BloscCompressor& compressor)
+  : compressor_{ compressor }
+{
+}
+
+zarr::BloscEncoder::~BloscEncoder() noexcept
+{
+    flush();
+    close_file();
+}
+
+size_t
+zarr::BloscEncoder::flush_impl()
+{
+    auto* buf_c = new uint8_t[cursor_ + BLOSC_MAX_OVERHEAD];
+    CHECK(buf_c);
+
+    const auto nbytes_out =
+      (size_t)blosc_compress_ctx(compressor_.clevel_,
+                                 compressor_.shuffle_,
+                                 bytes_per_pixel_,
+                                 cursor_,
+                                 buf_.data(),
+                                 buf_c,
+                                 cursor_ + BLOSC_MAX_OVERHEAD,
+                                 compressor_.codec_id_.c_str(),
+                                 0 /* blocksize - 0:automatic */,
+                                 (int)std::thread::hardware_concurrency());
+
+    CHECK(file_write(file_handle_, 0, buf_c, buf_c + nbytes_out));
+
+    delete[] buf_c;
+    return nbytes_out;
+}
+
+void
+zarr::BloscEncoder::open_file_impl()
+{
+}
 
 extern "C" struct Storage*
 compressed_zarr_zstd_init()
