+++ src/zarr.cpp
@@ -189,6 +189,501 @@ zarr_reserve_image_shape(Storage* self_, const ImageShape* shape) noexcept
 // zarr namespace implementations
 //
 
+zarr::Zarr::Zarr()
+  : dimension_separator_{ '/' }
+  , frame_count_{ 0 }
+  , pixel_scale_um_{ 1, 1 }
+  , max_bytes_per_chunk_{ 0 }
+  , tiles_per_chunk_{ 0 }
+  , image_shape_{ 0 }
+  , tile_shape_{ 0 }
+{
+    initialize_thread_pool_(1);
+}
+
+zarr::Zarr::Zarr(size_t nthreads)
+  : dimension_separator_{ '/' }
+  , frame_count_{ 0 }
+  , pixel_scale_um_{ 1, 1 }
+  , max_bytes_per_chunk_{ 0 }
+  , tiles_per_chunk_{ 0 }
+  , image_shape_{ 0 }
+  , tile_shape_{ 0 }
+{
+    nthreads = std::clamp(
+      nthreads, (size_t)1, (size_t)std::thread::hardware_concurrency());
+
+    initialize_thread_pool_(nthreads);
+}
+
+zarr::Zarr::Zarr(BloscCompressor&& compressor)
+  : dimension_separator_{ '/' }
+  , frame_count_{ 0 }
+  , pixel_scale_um_{ 1, 1 }
+  , max_bytes_per_chunk_{ 0 }
+  , tiles_per_chunk_{ 0 }
+  , image_shape_{ 0 }
+  , tile_shape_{ 0 }
+{
+    initialize_thread_pool_(1);
+    compressor_ = std::move(compressor);
+}
+
+zarr::Zarr::~Zarr()
+{
+    if (!stop())
+        LOGE("Failed to stop on destruct!");
+}
+
+void
+zarr::Zarr::set(const StorageProperties* props)
+{
+    using namespace acquire::sink::zarr;
+    CHECK(props);
+
+    StoragePropertyMetadata meta{};
+    get_meta(&meta);
+
+    // checks the directory exists and is writable
+    validate_props(props);
+    data_dir_ = as_path(*props).string();
+
+    if (props->external_metadata_json.str)
+        external_metadata_json_ = props->external_metadata_json.str;
+
+    pixel_scale_um_ = props->pixel_scale_um;
+
+    // chunking
+    set_chunking(props->chunking, meta.chunking);
+}
+
+void
+zarr::Zarr::get(StorageProperties* props) const
+{
+    CHECK(Device_Ok == storage_properties_set_filename(
+                         props, data_dir_.c_str(), data_dir_.size()));
+    CHECK(Device_Ok == storage_properties_set_external_metadata(
+                         props,
+                         external_metadata_json_.c_str(),
+                         external_metadata_json_.size()));
+    props->pixel_scale_um = pixel_scale_um_;
+}
+
+void
+zarr::Zarr::get_meta(StoragePropertyMetadata* meta) const
+{
+    CHECK(meta);
+    *meta = { .chunking = {
+                .supported = 1,
+                .max_bytes_per_chunk = { .writable = 1,
+                                         .low = (float)(16 << 20),
+                                         .high = (float)(1 << 31),
+                                         .type = PropertyType_FixedPrecision },
+              } };
+}
+
+void
+zarr::Zarr::start()
+{
+    frame_count_ = 0;
+    create_data_directory_();
+    write_zgroup_json_();
+    write_group_zattrs_json_();
+    write_zarray_json_();
+    write_external_metadata_json_();
+}
+
+int
+zarr::Zarr::stop() noexcept
+{
+    int is_ok = 1;
+
+    if (DeviceState_Running == state) {
+        state = DeviceState_Armed;
+        is_ok = 0;
+
+        try {
+            write_zarray_json_(); // must precede close of chunk file
+            while (!frame_ptrs_.empty()) {
+                TRACE("Cycling: %llu frames remaining", cycle_());
+                clock_sleep_ms(nullptr, 50.0);
+            }
+            recover_threads_();
+            finalize_thread_pool_();
+            clear_writers_();
+            is_ok = 1;
+        } catch (const std::exception& exc) {
+            LOGE("Exception: %s\n", exc.what());
+        } catch (...) {
+            LOGE("Exception: (unknown)");
+        }
+    }
+
+    return is_ok;
+}
+
+size_t
+zarr::Zarr::append(const VideoFrame* frames, size_t nbytes)
+{
+    if (0 == nbytes)
+        return nbytes;
+
+    using namespace acquire::sink::zarr;
+
+    const VideoFrame* cur = nullptr;
+    const auto* end = (const VideoFrame*)((uint8_t*)frames + nbytes);
+    auto next = [&]() -> const VideoFrame* {
+        const uint8_t* p = ((const uint8_t*)cur) + cur->bytes_of_frame;
+        return (const VideoFrame*)p;
+    };
+
+    for (cur = frames; cur < end; cur = next()) {
+        frame_ptrs_.push(
+          new TiledFrame(const_cast<VideoFrame*>(cur), tile_shape_));
+        auto* tiled_frame = frame_ptrs_.back();
+
+        // handle incoming image shape
+        validate_image_shapes_equal(image_shape_, cur->shape);
+
+        // push the new frame to our writers
+        for (auto& writer : writers_)
+            writer->push_frame(tiled_frame);
+
+        ++frame_count_;
+    }
+
+    TRACE("Cycling: %lu frames on queue", cycle_());
+
+    return nbytes;
+}
+
+void
+zarr::Zarr::reserve_image_shape(const ImageShape* shape)
+{
+    CHECK(shape);
+    image_shape_ = *shape;
+    allocate_writers_();
+
+    tiles_per_chunk_ =
+      get_tiles_per_chunk(image_shape_, tile_shape_, max_bytes_per_chunk_);
+}
+
+void
+zarr::Zarr::set_chunking(const ChunkingProps& props, const ChunkingMeta& meta)
+{
+    max_bytes_per_chunk_ = std::clamp(props.max_bytes_per_chunk,
+                                      (uint32_t)meta.max_bytes_per_chunk.low,
+                                      (uint32_t)meta.max_bytes_per_chunk.high);
+
+    uint32_t tile_width = props.tile.width;
+    if (tile_width == 0 ||
+        (image_shape_.dims.width && tile_width > image_shape_.dims.width)) {
+        LOGE("%s. Setting width to %u.",
+             tile_width == 0 ? "Tile width not specified"
+                             : "Specified roi width is too large",
+             image_shape_.dims.width);
+        tile_width = image_shape_.dims.width;
+    }
+
+    uint32_t tile_height = props.tile.height;
+    if (tile_height == 0 ||
+        (image_shape_.dims.height && tile_height > image_shape_.dims.height)) {
+        LOGE("%s. Setting height to %u.",
+             tile_height == 0 ? "Tile height not specified"
+                              : "Specified roi height is too large",
+             image_shape_.dims.height);
+        tile_height = image_shape_.dims.height;
+    }
+
+    uint32_t tile_planes = props.tile.planes;
+    if (tile_planes == 0 ||
+        (image_shape_.dims.planes && tile_planes > image_shape_.dims.planes)) {
+        LOGE("%s. Setting planes to %u.",
+             tile_planes == 0 ? "Tile planes not specified"
+                              : "Specified roi planes is too large",
+             image_shape_.dims.planes);
+        tile_planes = image_shape_.dims.planes;
+    }
+
+    tile_shape_ = {
+        .dims = {
+          .width = tile_width,
+          .height = tile_height,
+          .planes = tile_planes,
+        },
+        .frame_channels = {0}
+    };
+}
+
+void
+zarr::Zarr::initialize_thread_pool_(size_t nthreads)
+{
+    for (auto i = 0; i < nthreads; ++i) {
+        auto t = new thread_t;
+        thread_init(t);
+        thread_pool_.push(t);
+    }
+}
+
+void
+zarr::Zarr::finalize_thread_pool_()
+{
+    while (!thread_pool_.empty()) {
+        thread_t* t = thread_pool_.front();
+        thread_pool_.pop();
+        delete t;
+    }
+}
+
+void
+zarr::Zarr::create_data_directory_() const
+{
+    namespace fs = std::filesystem;
+    if (fs::exists(data_dir_)) {
+        std::error_code ec;
+        EXPECT(fs::remove_all(data_dir_, ec),
+               R"(Failed to remove folder for "%s": %s)",
+               data_dir_.c_str(),
+               ec.message().c_str());
+    }
+
+    EXPECT(fs::create_directory(data_dir_),
+           "Failed to create folder for \"%s\"",
+           data_dir_.c_str());
+}
+
+void
+zarr::Zarr::write_zarray_json_() const
+{
+    namespace fs = std::filesystem;
+    using namespace acquire::sink::zarr;
+    using json = nlohmann::json;
+
+    const auto frames_per_chunk = std::min(frame_count_, tiles_per_chunk_);
+
+    json zarray_attrs = {
+        { "zarr_format", 2 },
+        { "shape",
+          {
+            (uint64_t)frame_count_,
+            image_shape_.dims.channels,
+            image_shape_.dims.height,
+            image_shape_.dims.width,
+          } },
+        { "chunks",
+          {
+            (uint64_t)frames_per_chunk,
+            1,
+            tile_shape_.dims.height,
+            tile_shape_.dims.width,
+          } },
+        { "dtype", sample_type_to_dtype(image_shape_.type) },
+        { "fill_value", 0 },
+        { "order", "C" },
+        { "filters", nullptr },
+        { "dimension_separator", std::string(1, dimension_separator_) },
+    };
+
+    if (compressor_.has_value())
+        zarray_attrs["compressor"] = compressor_.value();
+    else
+        zarray_attrs["compressor"] = nullptr;
+
+    std::string zarray_path = (fs::path(data_dir_) / "0" / ".zarray").string();
+    write_string(zarray_path, zarray_attrs.dump());
+}
+
+void
+zarr::Zarr::write_external_metadata_json_() const
+{
+    namespace fs = std::filesystem;
+    using json = nlohmann::json;
+
+    std::string zattrs_path = (fs::path(data_dir_) / "0" / ".zattrs").string();
+    write_string(zattrs_path, external_metadata_json_);
+}
+
+void
+zarr::Zarr::write_group_zattrs_json_() const
+{
+    namespace fs = std::filesystem;
+    using json = nlohmann::json;
+
+    json zgroup_attrs;
+    zgroup_attrs["multiscales"] = json::array({ json::object() });
+    zgroup_attrs["multiscales"][0]["version"] = "0.4";
+    zgroup_attrs["multiscales"][0]["axes"] = {
+        {
+          { "name", "t" },
+          { "type", "time" },
+        },
+        {
+          { "name", "c" },
+          { "type", "channel" },
+        },
+        {
+          { "name", "y" },
+          { "type", "space" },
+          { "unit", "micrometer" },
+        },
+        {
+          { "name", "x" },
+          { "type", "space" },
+          { "unit", "micrometer" },
+        },
+    };
+    zgroup_attrs["multiscales"][0]["datasets"] = {
+        {
+          { "path", "0" },
+          { "coordinateTransformations",
+            {
+              {
+                { "type", "scale" },
+                { "scale", { 1, 1, pixel_scale_um_.y, pixel_scale_um_.x } },
+              },
+            } },
+        },
+    };
+
+    std::string zattrs_path = (fs::path(data_dir_) / ".zattrs").string();
+    write_string(zattrs_path, zgroup_attrs.dump(4));
+}
+
+void
+zarr::Zarr::write_zgroup_json_() const
+{
+    namespace fs = std::filesystem;
+    using json = nlohmann::json;
+
+    const json zgroup = { { "zarr_format", 2 } };
+    std::string zgroup_path = (fs::path(data_dir_) / ".zgroup").string();
+    write_string(zgroup_path, zgroup.dump());
+}
+
+void
+zarr::Zarr::allocate_writers_()
+{
+    auto frame_rois = make_frame_rois(image_shape_, tile_shape_);
+    CHECK(!frame_rois.empty());
+    TRACE("Allocating %llu writers", frame_rois.size());
+
+    std::for_each(
+      writers_.begin(), writers_.end(), [](ChunkWriter* w) { delete w; });
+    writers_.clear();
+    for (const auto& roi : frame_rois) {
+        BaseEncoder* encoder;
+        size_t buf_size;
+        if (compressor_.has_value()) {
+            buf_size = get_bytes_per_chunk(
+              image_shape_, tile_shape_, max_bytes_per_chunk_);
+            CHECK(encoder = new BloscEncoder(compressor_.value()));
+        } else {
+            buf_size = get_bytes_per_tile(image_shape_, tile_shape_);
+            CHECK(encoder = new RawEncoder());
+        }
+
+        encoder->allocate_buffer(buf_size);
+        encoder->set_bytes_per_pixel(bytes_per_sample_type(image_shape_.type));
+        auto writer = new ChunkWriter(roi, max_bytes_per_chunk_, encoder);
+        CHECK(writer);
+
+        writer->set_dimension_separator(dimension_separator_);
+        writer->set_base_directory(data_dir_);
+        writers_.push_back(writer);
+    }
+
+    while (writers_.size() < thread_pool_.size()) {
+        thread_t* t = thread_pool_.front();
+        thread_pool_.pop();
+        delete t;
+    }
+}
+
+void
+zarr::Zarr::clear_writers_()
+{
+    for (auto& writer : writers_) {
+        writer->close_current_file();
+        delete writer;
+    }
+    writers_.clear();
+}
+
+void
+zarr::Zarr::assign_threads_()
+{
+    // sort threads by number of frames still needing to be written
+    std::sort(
+      writers_.begin(), writers_.end(), [](ChunkWriter* a, ChunkWriter* b) {
+          return a->active_frames() > b->active_frames();
+      });
+
+    // thread_pool_ has at most as many threads as there are writers
+    while (!thread_pool_.empty()) {
+        thread_t* t = thread_pool_.front();
+        thread_pool_.pop();
+
+        while (t != nullptr) {
+            for (auto& writer : writers_) {
+                if (!writer->has_thread()) {
+                    writer->assign_thread(&t);
+                    break;
+                }
+            }
+        }
+    }
+}
+
+void
+zarr::Zarr::recover_threads_()
+{
+    for (auto& writer : writers_) {
+        thread_t* t = writer->release_thread();
+        if (nullptr != t)
+            thread_pool_.push(t);
+    }
+}
+
+/// make a single pass through the frame queue and check if any writers are
+/// still using any queued frames
+
+void
+zarr::Zarr::release_finished_frames_()
+{
+    const size_t nframes = frame_ptrs_.size();
+    for (auto i = 0; i < nframes; ++i) {
+        TiledFrame* frame = frame_ptrs_.front();
+        frame_ptrs_.pop();
+
+        bool replace_frame = false;
+        for (auto& writer : writers_) {
+            if (writer->has_frame(frame->frame_id())) {
+                replace_frame = true;
+                break;
+            }
+        }
+
+        if (replace_frame)
+            frame_ptrs_.push(frame);
+        else
+            delete frame;
+    }
+}
+
+size_t
+zarr::Zarr::cycle_()
+{
+    if (writers_.size() > thread_pool_.size())
+        recover_threads_();
+
+    release_finished_frames_();
+
+    if (!frame_ptrs_.empty())
+        assign_threads_();
+
+    return frame_ptrs_.size();
+}
+
 /// \brief Check that the StorageProperties are valid.
 /// \details Assumes either an empty or valid JSON metadata string and a
 /// filename string that points to a writable directory. \param props Storage
@@ -427,7 +922,7 @@ zarr_init()
 {
     try {
         auto nthreads = (int)std::thread::hardware_concurrency();
-        return new zarr::Zarr<zarr::RawEncoder>(std::max(nthreads - 2, 1));
+        return new zarr::Zarr(std::max(nthreads - 3, 1));
     } catch (const std::exception& exc) {
         LOGE("Exception: %s\n", exc.what());
     } catch (...) {
